+++ ReleaseNotes/release_notes.md
@@ -1,6 +1,9 @@
 ## New features
- 1. add `-v/--version` flag for `ostap`  : print the version and exit
     
+ 1. add `-v/--version` flag for `ostap`  : print the version and exit
+ 1. refactor the main ostap parser into separate module  (it speeds up 'soatp -v' action)  
+
+       
 ## Backward-imcompatible 
 
 ## Bug fixes